backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# PDF Processing
pdfplumber==0.11.4

# LLM Providers
google-generativeai==0.8.4
groq==0.15.0

# RAG Components & Evaluation
chromadb==0.6.3
sentence-transformers==3.3.1
rank-bm25==0.2.2
ragas>=0.2.14
datasets>=3.3.2
langchain-google-genai>=2.0.10

# Data Validation & Settings
pydantic==2.10.4
pydantic-settings==2.7.1

# Privacy & PII Scrubbing
presidio-analyzer>=2.2.35
presidio-anonymizer>=2.2.35
spacy>=3.7.4

# Database
aiosqlite==0.20.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.3

# Testing
pytest==8.3.4
httpx==0.28.1
pytest-asyncio==0.25.0
